File DND to other apps
Closes https://github.com/lxqt/lxqt-archiver/issues/79

This is what can be done with purely Qt methods for file DND from lxqt-archiver to other apps. It supports multiple files/folders.

Its advantage is that (unlike in engrampa) dropping can happen on any app that accepts a file drop (like a text editor or an image viewer), not just on those supporting the Direct Save Protocol (XDS). As a result, for example, if a file is dropped on Desktop, it'll be positioned at the drop point.

Its drawback may be that, with huge archives, the drop (cursor) will not be ready immediately because the dragged files should be first extracted in `/tmp`. However, the progress-bar informs the user when the drop is ready. So, I think we shouldn't deprive ourselves of this DND.