Prevent dropping into self
When an item is dragged from an lxqt-archiver window, usually it cannot be dropped into it again because only archives can be dropped into lxqt-archiver windows. But some archives contain other archives (like deb packages). This patch guarantees that the items of a window are never dropped into it — although they can be dropped into another window.