Added milliseconds to name of temporary extraction folder
The temporary extraction folder is used for DND and viewing items. Since its name is determined when the archive is loaded and because multiple archives can be opened almost simultaneously, milliseconds should be added to its name.

Also the context of a lambda connection is fixed for its appropriate disconnection.