Added dedicated support for `bsdtar` (#417)
Previously, `bsdtar` was used for `rpm` and compressed disk images in two different source files. Now it's in a single source file (`fr-command-bsdtar.c`), such that more archive formats could be added to it easily if needed.

Also, an explicit support for Android packages is added by using `7z` and `bsdtar`, in this order.

NOTE: The old source files are kept only for comparison, without being used.